History hour attribute
**After review update db structure on server**
Now the hour attribute from history is a real number with 2 decimal places
and maximum of 9999,99
Every insert will be rounded to 2 decimal places

+++ src/main/resources/scripts/create_tables.sql
@@ -1,10 +1,10 @@
 CREATE TABLE Users (
   id        SERIAL PRIMARY KEY,
-  mail      VARCHAR (63) NOT NULL,
-  admin     BOOLEAN      NOT NULL,
-  active    BOOLEAN      NOT NULL,
-  lang      VARCHAR(2)   NOT NULL    DEFAULT 'pl',
-  work_time REAL                     DEFAULT 8.0
+  mail      VARCHAR(63) NOT NULL,
+  admin     BOOLEAN     NOT NULL,
+  active    BOOLEAN     NOT NULL,
+  lang      VARCHAR(2)  NOT NULL    DEFAULT 'pl',
+  work_time REAL                    DEFAULT 8.0
 );
 
 CREATE UNIQUE INDEX users_mail_index
@@ -17,20 +17,20 @@ CREATE TABLE Requests (
   requester_id INT REFERENCES Users (id) NOT NULL,
   start_date   DATE                      NOT NULL,
   end_date     DATE                      NOT NULL,
-  type         VARCHAR (25)              NOT NULL   DEFAULT 'NORMAL',
-  type_info    VARCHAR (25),
-  status       VARCHAR (255)             NOT NULL   DEFAULT 'PENDING'
+  type         VARCHAR(25)               NOT NULL   DEFAULT 'NORMAL',
+  type_info    VARCHAR(25),
+  status       VARCHAR(255)              NOT NULL   DEFAULT 'PENDING'
 );
 
 CREATE INDEX requests_modified_index
   ON Requests (modified);
 
 CREATE TABLE Acceptances (
-  id          SERIAL PRIMARY KEY,
-  request_id  INT REFERENCES Requests (id)   NOT NULL,
-  leader_id   INT REFERENCES Users    (id)   NOT NULL,
-  decider_id  INT REFERENCES Users    (id),
-  accepted    BOOLEAN                        NOT NULL   DEFAULT FALSE
+  id         SERIAL PRIMARY KEY,
+  request_id INT REFERENCES Requests (id)   NOT NULL,
+  leader_id  INT REFERENCES Users (id)      NOT NULL,
+  decider_id INT REFERENCES Users (id),
+  accepted   BOOLEAN                        NOT NULL   DEFAULT FALSE
 );
 
 CREATE INDEX acceptances_requestId_index
@@ -41,17 +41,17 @@ CREATE TABLE History (
   created    TIMESTAMP                    NOT NULL,
   user_id    INT REFERENCES Users (id)    NOT NULL,
   decider_id INT REFERENCES Users (id),
-  request_id INT REFERENCES Requests(id),
-  hours      REAL                          NOT NULL,
+  request_id INT REFERENCES Requests (id),
+  hours      DECIMAL(6, 2)                NOT NULL,
   comment    VARCHAR(255)
 );
 
 -- CREATE UNIQUE INDEX history_userId_index ON History (user_id);
 
 CREATE TABLE Holidays (
-  id    SERIAL         PRIMARY KEY,
-  name  VARCHAR (100)  NOT NULL,
-  date  DATE           NOT NULL
+  id   SERIAL PRIMARY KEY,
+  name VARCHAR(100) NOT NULL,
+  date DATE         NOT NULL
 );
 
 ALTER SEQUENCE users_id_seq RESTART;
